requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
image==1.5.33
imageio==2.36.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.25.0
scipy==1.15.0
six==1.17.0
sqlparse==0.5.3
tifffile==2024.12.12
uritemplate==4.1.1
urllib3==2.3.0
